requirements.txt
# CS4241 RAG — Kofi Assan (1002300129)
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
